Struct diem_types::account_state::AccountState

pub struct AccountState(_);

Implementations

impl AccountState

pub fn get_account_address(&self) -> Result<Option<AccountAddress>>

pub fn get_account_resource(&self) -> Result<Option<AccountResource>>

pub fn get_crsn_resource(&self) -> Result<Option<CRSNResource>>

pub fn get_balance_resources(
&self
) -> Result<BTreeMap<Identifier, BalanceResource>>

pub fn get_preburn_balances(
&self
) -> Result<BTreeMap<Identifier, PreburnResource>>

pub fn get_preburn_queue_balances(
&self
) -> Result<BTreeMap<Identifier, PreburnQueueResource>>

pub fn get_chain_id_resource(&self) -> Result<Option<ChainIdResource>>

pub fn get_configuration_resource(
&self
) -> Result<Option<ConfigurationResource>>

pub fn get_diem_timestamp_resource(
&self
) -> Result<Option<DiemTimestampResource>>

pub fn get_validator_config_resource(
&self
) -> Result<Option<ValidatorConfigResource>>

pub fn get_validator_operator_config_resource(
&self
) -> Result<Option<ValidatorOperatorConfigResource>>

pub fn get_freezing_bit(&self) -> Result<Option<FreezingBit>>

pub fn get_account_role(&self) -> Result<Option<AccountRole>>

pub fn get_validator_set(&self) -> Result<Option<ValidatorSet>>

pub fn get_diem_version(&self) -> Result<Option<DiemVersion>>

pub fn get_vm_publishing_option(&self) -> Result<Option<VMPublishingOption>>

pub fn get_registered_currency_info_resources(
&self
) -> Result<Vec<CurrencyInfoResource>>

pub fn get_diem_block_resource(&self) -> Result<Option<DiemBlockResource>>

pub fn get(&self, key: &[u8]) -> Option<&Vec<u8>>

pub fn get_resource_impl<T: DeserializeOwned>(
&self,
key: &[u8]
) -> Result<Option<T>>

pub fn insert(&mut self, key: Vec<u8>, value: Vec<u8>) -> Option<Vec<u8>>

pub fn remove(&mut self, key: &[u8]) -> Option<Vec<u8>>

pub fn iter(&self) -> impl Iterator<Item = (&Vec<u8>, &Vec<u8>)>

pub fn get_config<T: OnChainConfig>(&self) -> Result<Option<T>>

pub fn get_resource<T: MoveResource>(&self) -> Result<Option<T>>

pub fn get_modules(&self) -> impl Iterator<Item = &Vec<u8>>

Return an iterator over the module values stored under this account

pub fn get_resources(&self) -> impl Iterator<Item = (StructTag, &[u8])>

Return an iterator over all resources stored under this account.

Note that resource access Paths that fail to deserialize will be silently ignored.

pub fn get_resources_with_type<T: MoveResource>(
&self
) -> impl Iterator<Item = Result<(StructTag, T)>> + '_

Given a particular MoveResource, return an iterator with all instances of that resource (there may be multiple with different generic type parameters).