diem_types

Module block_info

Structs

BlockInfo

This structure contains all the information needed for tracking a block without having access to the block or its execution output state. It assumes that the block is the last block executed within the ledger.

PivotBlockDecision

PivotBlockDecisionHasher

Cryptographic hasher for an BCS-serializable #item

Constants

GENESIS_EPOCH

GENESIS_ROUND

GENESIS_TIMESTAMP_USECS

GENESIS_VERSION

Type Aliases

Round

The round of a block is a consensus-internal counter, which starts with 0 and increases monotonically.

View