diem_types

Module epoch_change

Structs

EpochChangeProof

A vector of LedgerInfo with contiguous increasing epoch numbers to prove a sequence of epoch changes from the first LedgerInfo’s epoch.

Traits

Verifier

The verification of the epoch change proof starts with verifier that is trusted by the client: could be either a waypoint (upon startup) or a known epoch info.