Struct diem_types::proof::accumulator::InMemoryAccumulator

pub struct InMemoryAccumulator<H> { /* fields omitted */ }

The Accumulator implementation.

Implementations

impl<H> InMemoryAccumulator<H> where
H: CryptoHasher,

pub fn new(
frozen_subtree_roots: Vec<HashValue>,
num_leaves: LeafCount
) -> Result<Self>

Constructs a new accumulator with roots of existing frozen subtrees. Returns error if the number of frozen subtree roots does not match the number of leaves.

pub fn from_leaves(leaves: &[HashValue]) -> Self

Constructs a new accumulator with given leaves.

pub fn append(&self, leaves: &[HashValue]) -> Self

Appends a list of new leaves to an existing accumulator. Since the accumulator is immutable, the existing one remains unchanged and a new one representing the result is returned.

pub fn append_subtrees(
&self,
subtrees: &[HashValue],
num_new_leaves: LeafCount
) -> Result<Self>

Appends a list of new subtrees to the existing accumulator. This is similar to append except that the new leaves themselves are not known and they are represented by subtrees. As an example, given the following accumulator that currently has 10 leaves, the frozen subtree roots and the new subtrees are annotated below. Note that in this case subtrees[0] represents two new leaves A and B, subtrees[1] represents four new leaves C, D, E and F, subtrees[2] represents four new leaves G, H, I and J, and the last subtrees[3] represents one new leaf K.

                                                                          new_root
                                                                        /          \
                                                                      /              \
                                                                    /                  \
                                                                  /                      \
                                                                /                          \
                                                              /                              \
                                                            /                                  \
                                                          /                                      \
                                                        /                                          \
                                                      /                                              \
                                                    /                                                  \
                                                  /                                                      \
                                                /                                                          \
                                        old_root                                                            o
                                       /        \                                                          / \
                                     /            \                                                       /   placeholder
                                   /                \                                                    /
                                 /                    \                                                 /
                               /                        \                                              /
                             /                            \                                           o
                           /                                \                                        / \
                         /                                    \                                     /    \
                       /                                       o                                  /        \
frozen_subtree_roots[0]                                      /   \                              /            \
                   /   \                                    /     \                           /                \
                  /     \                                  /       \                        /                    \
                 o       o                                o         subtrees[1]  subtrees[2]                     o
                / \     / \                              / \                / \          / \                    / \
               o   o   o   o      frozen_subtree_roots[1]   subtrees[0]    o   o        o   o                  o   placeholder
              / \ / \ / \ / \                         / \           / \   / \ / \      / \ / \                / \
              o o o o o o o o                         o o           A B   C D E F      G H I J  K (subtrees[3]) placeholder

pub fn root_hash(&self) -> HashValue

Returns the root hash of the accumulator.

pub fn version(&self) -> u64

pub fn frozen_subtree_roots(&self) -> &Vec<HashValue>

Returns the set of frozen subtree roots in this accumulator

pub fn num_leaves(&self) -> LeafCount

Returns the total number of leaves in this accumulator.

pub fn is_empty(&self) -> bool

Returns true if this accumulator is empty and has no leaves.