diem_types::transaction

Module metadata

This file implements Diem transaction metadata types to allow easy parsing and introspection into metadata, whether the transaction is using regular subaddressing, is subject to travel rule or corresponds to an on-chain payment refund.

Structs

CoinTradeMetadataV0

Transaction metadata format for coin trades (purchases/sells)

GeneralMetadataV0

Transaction metadata for regular addressing with optional subaddressing or refunded transaction reference

RefundMetadataV0

Transaction metadata format for transactions subject to refund transaction

TravelRuleMetadataV0

Transaction metadata format for transactions subject to travel rule

UnstructuredBytesMetadata

Opaque binary transaction metadata

Enums

CoinTradeMetadata

List of supported transaction metadata format versions for coin trade transaction

GeneralMetadata

List of supported transaction metadata format versions for regular addressing with optional subaddressing or refund reference

Metadata

List of all supported metadata types

RefundMetadata

List of supported transaction metadata format versions for refund transaction

RefundReason

TravelRuleMetadata

List of supported transaction metadata format versions for transactions subject to travel rule