Trait diem_types::transaction::TransactionInfoTrait

pub trait TransactionInfoTrait: Clone + CryptoHash + Debug + Display + Eq + Serialize + DeserializeOwned {
    fn new(
        transaction_hash: HashValue, 
        state_root_hash: HashValue, 
        event_root_hash: HashValue, 
        gas_used: u64, 
        status: KeptVMStatus
    ) -> Self;

    fn transaction_hash(&self) -> HashValue;

    fn state_root_hash(&self) -> HashValue;

    fn event_root_hash(&self) -> HashValue;

    fn gas_used(&self) -> u64;

    fn status(&self) -> &KeptVMStatus;
}

Required methods

fn new(
transaction_hash: HashValue,
state_root_hash: HashValue,
event_root_hash: HashValue,
gas_used: u64,
status: KeptVMStatus
) -> Self

Constructs a new TransactionInfo object using transaction hash, state root hash and event root hash.

fn transaction_hash(&self) -> HashValue

Returns the hash of this transaction.

fn state_root_hash(&self) -> HashValue

Returns root hash of Sparse Merkle Tree describing the world state at the end of this transaction.

fn event_root_hash(&self) -> HashValue

Returns the root hash of Merkle Accumulator storing all events emitted during this transaction.

fn gas_used(&self) -> u64

Returns the amount of gas used by this transaction.

fn status(&self) -> &KeptVMStatus

Resturns the Status from the VM for this transaction.