diem_types::validator_verifier

Struct ValidatorVerifier

pub struct ValidatorVerifier { /* private fields */ }

Supports validation of signatures for known authors with individual voting powers. This struct can be used for all signature verification operations including block and network signature verification, respectively.

Implementations

impl ValidatorVerifier

pub fn new( address_to_validator_info: BTreeMap<AccountAddress, ValidatorConsensusInfo>, ) -> Self

Initialize with a map of account address to validator info and set quorum size to default (2f + 1) or zero if address_to_validator_info is empty.

pub fn new_single( author: AccountAddress, public_key: ConsensusPublicKey, vrf_public_key: Option<ConsensusVRFPublicKey>, ) -> Self

Helper method to initialize with a single author and public key with quorum voting power 1.

pub fn verify<T: Serialize + CryptoHash>( &self, author: AccountAddress, message: &T, signature: &ConsensusSignature, ) -> Result<(), VerifyError>

Verify the correctness of a signature of a message by a known author.

pub fn verify_vrf( &self, author: AccountAddress, seed: &[u8], proof: &ConsensusVRFProof, ) -> Result<(), VerifyError>

Verify the correctness of a signature of a message by a known author.

pub fn verify_aggregated_struct_signature<T: CryptoHash + Serialize>( &self, message: &T, aggregated_signature: &BTreeMap<AccountAddress, ConsensusSignature>, ) -> Result<(), VerifyError>

This function will successfully return when at least quorum_size signatures of known authors are successfully verified. Also, an aggregated signature is considered invalid if any of the attached signatures is invalid or it does not correspond to a known author. The latter is to prevent malicious users from adding arbitrary content to the signature payload that would go unnoticed.

pub fn batch_verify_aggregated_signatures<T: CryptoHash + Serialize>( &self, message: &T, aggregated_signature: &BTreeMap<AccountAddress, ConsensusSignature>, ) -> Result<(), VerifyError>

This function will try batch signature verification and falls back to normal iterated verification if batching fails.

pub fn check_voting_power<'a>( &self, authors: impl Iterator<Item = &'a AccountAddress>, ) -> Result<(), VerifyError>

Ensure there is at least quorum_voting_power in the provided signatures and there are only known authors. According to the threshold verification policy, invalid public keys are not allowed.

pub fn get_public_key( &self, author: &AccountAddress, ) -> Option<ConsensusPublicKey>

Returns the public key for this address.

pub fn get_vrf_public_key( &self, author: &AccountAddress, ) -> Option<Option<ConsensusVRFPublicKey>>

Returns the VRF public key for this address.

pub fn get_voting_power(&self, author: &AccountAddress) -> Option<u64>

Returns the voting power for this address.

pub fn get_ordered_account_addresses_iter( &self, ) -> impl Iterator<Item = AccountAddress> + '_

Returns an ordered list of account addresses as an Iterator.

pub fn len(&self) -> usize

Returns the number of authors to be validated.

pub fn is_empty(&self) -> bool

Is there at least one author?

pub fn quorum_voting_power(&self) -> u64

Returns quorum voting power.

pub fn total_voting_power(&self) -> u64

pub fn address_to_validator_info( &self, ) -> &BTreeMap<AccountAddress, ValidatorConsensusInfo>

pub fn extra_vote_count<'a>( &self, signers: impl Iterator<Item = &'a AccountAddress>, ) -> Result<u64>

Return the extra vote compared to quorum_voting_power.