diem_types::vm_status

Enum StatusCode

#[repr(u64)]
pub enum StatusCode {

    UNKNOWN_VALIDATION_STATUS = 0,
    INVALID_SIGNATURE = 1,
    INVALID_AUTH_KEY = 2,
    SEQUENCE_NUMBER_TOO_OLD = 3,
    SEQUENCE_NUMBER_TOO_NEW = 4,
    INSUFFICIENT_BALANCE_FOR_TRANSACTION_FEE = 5,
    TRANSACTION_EXPIRED = 6,
    SENDING_ACCOUNT_DOES_NOT_EXIST = 7,
    REJECTED_WRITE_SET = 8,
    INVALID_WRITE_SET = 9,
    EXCEEDED_MAX_TRANSACTION_SIZE = 10,
    UNKNOWN_SCRIPT = 11,
    UNKNOWN_MODULE = 12,
    MAX_GAS_UNITS_EXCEEDS_MAX_GAS_UNITS_BOUND = 13,
    MAX_GAS_UNITS_BELOW_MIN_TRANSACTION_GAS_UNITS = 14,
    GAS_UNIT_PRICE_BELOW_MIN_BOUND = 15,
    GAS_UNIT_PRICE_ABOVE_MAX_BOUND = 16,
    INVALID_GAS_SPECIFIER = 17,
    SENDING_ACCOUNT_FROZEN = 18,
    UNABLE_TO_DESERIALIZE_ACCOUNT = 19,
    CURRENCY_INFO_DOES_NOT_EXIST = 20,
    INVALID_MODULE_PUBLISHER = 21,
    NO_ACCOUNT_ROLE = 22,
    BAD_CHAIN_ID = 23,
    SEQUENCE_NUMBER_TOO_BIG = 24,
    BAD_TRANSACTION_FEE_CURRENCY = 25,
    FEATURE_UNDER_GATING = 26,
    CFX_UNEXPECTED_TX = 27,
    CFX_INVALID_TX = 28,
    ELECTION_NON_EXISITENT_NODE = 29,
    ELECTION_NON_ACCEPTED_NODE = 30,
    ELECTION_TERGET_TERM_NOT_OPEN = 31,
    ELECTION_WITHOUT_VOTES = 32,
    PIVOT_DECISION_HEIGHT_TOO_OLD = 33,
    UNKNOWN_VERIFICATION_ERROR = 1_000,
    INDEX_OUT_OF_BOUNDS = 1_001,
    INVALID_SIGNATURE_TOKEN = 1_003,
    RECURSIVE_STRUCT_DEFINITION = 1_005,
    FIELD_MISSING_TYPE_ABILITY = 1_006,
    INVALID_FALL_THROUGH = 1_007,
    NEGATIVE_STACK_SIZE_WITHIN_BLOCK = 1_009,
    INVALID_MAIN_FUNCTION_SIGNATURE = 1_011,
    DUPLICATE_ELEMENT = 1_012,
    INVALID_MODULE_HANDLE = 1_013,
    UNIMPLEMENTED_HANDLE = 1_014,
    LOOKUP_FAILED = 1_017,
    TYPE_MISMATCH = 1_020,
    MISSING_DEPENDENCY = 1_021,
    POP_WITHOUT_DROP_ABILITY = 1_023,
    BR_TYPE_MISMATCH_ERROR = 1_025,
    ABORT_TYPE_MISMATCH_ERROR = 1_026,
    STLOC_TYPE_MISMATCH_ERROR = 1_027,
    STLOC_UNSAFE_TO_DESTROY_ERROR = 1_028,
    UNSAFE_RET_LOCAL_OR_RESOURCE_STILL_BORROWED = 1_029,
    RET_TYPE_MISMATCH_ERROR = 1_030,
    RET_BORROWED_MUTABLE_REFERENCE_ERROR = 1_031,
    FREEZEREF_TYPE_MISMATCH_ERROR = 1_032,
    FREEZEREF_EXISTS_MUTABLE_BORROW_ERROR = 1_033,
    BORROWFIELD_TYPE_MISMATCH_ERROR = 1_034,
    BORROWFIELD_BAD_FIELD_ERROR = 1_035,
    BORROWFIELD_EXISTS_MUTABLE_BORROW_ERROR = 1_036,
    COPYLOC_UNAVAILABLE_ERROR = 1_037,
    COPYLOC_WITHOUT_COPY_ABILITY = 1_038,
    COPYLOC_EXISTS_BORROW_ERROR = 1_039,
    MOVELOC_UNAVAILABLE_ERROR = 1_040,
    MOVELOC_EXISTS_BORROW_ERROR = 1_041,
    BORROWLOC_REFERENCE_ERROR = 1_042,
    BORROWLOC_UNAVAILABLE_ERROR = 1_043,
    BORROWLOC_EXISTS_BORROW_ERROR = 1_044,
    CALL_TYPE_MISMATCH_ERROR = 1_045,
    CALL_BORROWED_MUTABLE_REFERENCE_ERROR = 1_046,
    PACK_TYPE_MISMATCH_ERROR = 1_047,
    UNPACK_TYPE_MISMATCH_ERROR = 1_048,
    READREF_TYPE_MISMATCH_ERROR = 1_049,
    READREF_WITHOUT_COPY_ABILITY = 1_050,
    READREF_EXISTS_MUTABLE_BORROW_ERROR = 1_051,
    WRITEREF_TYPE_MISMATCH_ERROR = 1_052,
    WRITEREF_WITHOUT_DROP_ABILITY = 1_053,
    WRITEREF_EXISTS_BORROW_ERROR = 1_054,
    WRITEREF_NO_MUTABLE_REFERENCE_ERROR = 1_055,
    INTEGER_OP_TYPE_MISMATCH_ERROR = 1_056,
    BOOLEAN_OP_TYPE_MISMATCH_ERROR = 1_057,
    EQUALITY_OP_TYPE_MISMATCH_ERROR = 1_058,
    EXISTS_WITHOUT_KEY_ABILITY_OR_BAD_ARGUMENT = 1_059,
    BORROWGLOBAL_TYPE_MISMATCH_ERROR = 1_060,
    BORROWGLOBAL_WITHOUT_KEY_ABILITY = 1_061,
    MOVEFROM_TYPE_MISMATCH_ERROR = 1_062,
    MOVEFROM_WITHOUT_KEY_ABILITY = 1_063,
    MOVETO_TYPE_MISMATCH_ERROR = 1_064,
    MOVETO_WITHOUT_KEY_ABILITY = 1_065,
    MODULE_ADDRESS_DOES_NOT_MATCH_SENDER = 1_067,
    NO_MODULE_HANDLES = 1_068,
    POSITIVE_STACK_SIZE_AT_BLOCK_END = 1_069,
    MISSING_ACQUIRES_ANNOTATION = 1_070,
    EXTRANEOUS_ACQUIRES_ANNOTATION = 1_071,
    DUPLICATE_ACQUIRES_ANNOTATION = 1_072,
    INVALID_ACQUIRES_ANNOTATION = 1_073,
    GLOBAL_REFERENCE_ERROR = 1_074,
    CONSTRAINT_NOT_SATISFIED = 1_075,
    NUMBER_OF_TYPE_ARGUMENTS_MISMATCH = 1_076,
    LOOP_IN_INSTANTIATION_GRAPH = 1_077,
    ZERO_SIZED_STRUCT = 1_080,
    LINKER_ERROR = 1_081,
    INVALID_CONSTANT_TYPE = 1_082,
    MALFORMED_CONSTANT_DATA = 1_083,
    EMPTY_CODE_UNIT = 1_084,
    INVALID_LOOP_SPLIT = 1_085,
    INVALID_LOOP_BREAK = 1_086,
    INVALID_LOOP_CONTINUE = 1_087,
    UNSAFE_RET_UNUSED_VALUES_WITHOUT_DROP = 1_088,
    TOO_MANY_LOCALS = 1_089,
    GENERIC_MEMBER_OPCODE_MISMATCH = 1_090,
    FUNCTION_RESOLUTION_FAILURE = 1_091,
    INVALID_OPERATION_IN_SCRIPT = 1_094,
    DUPLICATE_MODULE_NAME = 1_095,
    BACKWARD_INCOMPATIBLE_MODULE_UPDATE = 1_096,
    CYCLIC_MODULE_DEPENDENCY = 1_097,
    NUMBER_OF_ARGUMENTS_MISMATCH = 1_098,
    INVALID_PARAM_TYPE_FOR_DESERIALIZATION = 1_099,
    FAILED_TO_DESERIALIZE_ARGUMENT = 1_100,
    NUMBER_OF_SIGNER_ARGUMENTS_MISMATCH = 1_101,
    CALLED_SCRIPT_VISIBLE_FROM_NON_SCRIPT_VISIBLE = 1_102,
    EXECUTE_SCRIPT_FUNCTION_CALLED_ON_NON_SCRIPT_VISIBLE = 1_103,
    INVALID_FRIEND_DECL_WITH_SELF = 1_104,
    INVALID_FRIEND_DECL_WITH_MODULES_OUTSIDE_ACCOUNT_ADDRESS = 1_105,
    INVALID_FRIEND_DECL_WITH_MODULES_IN_DEPENDENCIES = 1_106,
    CYCLIC_MODULE_FRIENDSHIP = 1_107,
    UNKNOWN_INVARIANT_VIOLATION_ERROR = 2_000,
    EMPTY_VALUE_STACK = 2_003,
    PC_OVERFLOW = 2_005,
    VERIFICATION_ERROR = 2_006,
    STORAGE_ERROR = 2_008,
    INTERNAL_TYPE_ERROR = 2_009,
    EVENT_KEY_MISMATCH = 2_010,
    UNREACHABLE = 2_011,
    VM_STARTUP_FAILURE = 2_012,
    UNEXPECTED_ERROR_FROM_KNOWN_MOVE_FUNCTION = 2_015,
    VERIFIER_INVARIANT_VIOLATION = 2_016,
    UNEXPECTED_VERIFIER_ERROR = 2_017,
    UNEXPECTED_DESERIALIZATION_ERROR = 2_018,
    FAILED_TO_SERIALIZE_WRITE_SET_CHANGES = 2_019,
    FAILED_TO_DESERIALIZE_RESOURCE = 2_020,
    TYPE_RESOLUTION_FAILURE = 2_021,
    UNKNOWN_BINARY_ERROR = 3_000,
    MALFORMED = 3_001,
    BAD_MAGIC = 3_002,
    UNKNOWN_VERSION = 3_003,
    UNKNOWN_TABLE_TYPE = 3_004,
    UNKNOWN_SIGNATURE_TYPE = 3_005,
    UNKNOWN_SERIALIZED_TYPE = 3_006,
    UNKNOWN_OPCODE = 3_007,
    BAD_HEADER_TABLE = 3_008,
    UNEXPECTED_SIGNATURE_TYPE = 3_009,
    DUPLICATE_TABLE = 3_010,
    UNKNOWN_ABILITY = 3_013,
    UNKNOWN_NATIVE_STRUCT_FLAG = 3_014,
    BAD_U64 = 3_019,
    BAD_U128 = 3_020,
    VALUE_SERIALIZATION_ERROR = 3_022,
    VALUE_DESERIALIZATION_ERROR = 3_023,
    CODE_DESERIALIZATION_ERROR = 3_024,
    INVALID_FLAG_BITS = 3_025,
    UNKNOWN_RUNTIME_STATUS = 4_000,
    EXECUTED = 4_001,
    OUT_OF_GAS = 4_002,
    RESOURCE_DOES_NOT_EXIST = 4_003,
    RESOURCE_ALREADY_EXISTS = 4_004,
    MISSING_DATA = 4_008,
    DATA_FORMAT_ERROR = 4_009,
    ABORTED = 4_016,
    ARITHMETIC_ERROR = 4_017,
    EXECUTION_STACK_OVERFLOW = 4_020,
    CALL_STACK_OVERFLOW = 4_021,
    VM_MAX_TYPE_DEPTH_REACHED = 4_024,
    VM_MAX_VALUE_DEPTH_REACHED = 4_025,
    UNKNOWN_STATUS = 18_446_744_073_709_551_615,

}

We don’t derive Arbitrary on this enum because it is too large and breaks proptest. It is written for a subset of these in proptest_types. We test conversion between this and protobuf with a hand-written test.

Variants

UNKNOWN_VALIDATION_STATUS = 0

INVALID_SIGNATURE = 1

INVALID_AUTH_KEY = 2

SEQUENCE_NUMBER_TOO_OLD = 3

SEQUENCE_NUMBER_TOO_NEW = 4

INSUFFICIENT_BALANCE_FOR_TRANSACTION_FEE = 5

TRANSACTION_EXPIRED = 6

SENDING_ACCOUNT_DOES_NOT_EXIST = 7

REJECTED_WRITE_SET = 8

INVALID_WRITE_SET = 9

EXCEEDED_MAX_TRANSACTION_SIZE = 10

UNKNOWN_SCRIPT = 11

UNKNOWN_MODULE = 12

MAX_GAS_UNITS_EXCEEDS_MAX_GAS_UNITS_BOUND = 13

MAX_GAS_UNITS_BELOW_MIN_TRANSACTION_GAS_UNITS = 14

GAS_UNIT_PRICE_BELOW_MIN_BOUND = 15

GAS_UNIT_PRICE_ABOVE_MAX_BOUND = 16

INVALID_GAS_SPECIFIER = 17

SENDING_ACCOUNT_FROZEN = 18

UNABLE_TO_DESERIALIZE_ACCOUNT = 19

CURRENCY_INFO_DOES_NOT_EXIST = 20

INVALID_MODULE_PUBLISHER = 21

NO_ACCOUNT_ROLE = 22

BAD_CHAIN_ID = 23

SEQUENCE_NUMBER_TOO_BIG = 24

BAD_TRANSACTION_FEE_CURRENCY = 25

FEATURE_UNDER_GATING = 26

CFX_UNEXPECTED_TX = 27

CFX_INVALID_TX = 28

ELECTION_NON_EXISITENT_NODE = 29

ELECTION_NON_ACCEPTED_NODE = 30

ELECTION_TERGET_TERM_NOT_OPEN = 31

ELECTION_WITHOUT_VOTES = 32

PIVOT_DECISION_HEIGHT_TOO_OLD = 33

UNKNOWN_VERIFICATION_ERROR = 1_000

INDEX_OUT_OF_BOUNDS = 1_001

INVALID_SIGNATURE_TOKEN = 1_003

RECURSIVE_STRUCT_DEFINITION = 1_005

FIELD_MISSING_TYPE_ABILITY = 1_006

INVALID_FALL_THROUGH = 1_007

NEGATIVE_STACK_SIZE_WITHIN_BLOCK = 1_009

INVALID_MAIN_FUNCTION_SIGNATURE = 1_011

DUPLICATE_ELEMENT = 1_012

INVALID_MODULE_HANDLE = 1_013

UNIMPLEMENTED_HANDLE = 1_014

LOOKUP_FAILED = 1_017

TYPE_MISMATCH = 1_020

MISSING_DEPENDENCY = 1_021

POP_WITHOUT_DROP_ABILITY = 1_023

BR_TYPE_MISMATCH_ERROR = 1_025

ABORT_TYPE_MISMATCH_ERROR = 1_026

STLOC_TYPE_MISMATCH_ERROR = 1_027

STLOC_UNSAFE_TO_DESTROY_ERROR = 1_028

UNSAFE_RET_LOCAL_OR_RESOURCE_STILL_BORROWED = 1_029

RET_TYPE_MISMATCH_ERROR = 1_030

RET_BORROWED_MUTABLE_REFERENCE_ERROR = 1_031

FREEZEREF_TYPE_MISMATCH_ERROR = 1_032

FREEZEREF_EXISTS_MUTABLE_BORROW_ERROR = 1_033

BORROWFIELD_TYPE_MISMATCH_ERROR = 1_034

BORROWFIELD_BAD_FIELD_ERROR = 1_035

BORROWFIELD_EXISTS_MUTABLE_BORROW_ERROR = 1_036

COPYLOC_UNAVAILABLE_ERROR = 1_037

COPYLOC_WITHOUT_COPY_ABILITY = 1_038

COPYLOC_EXISTS_BORROW_ERROR = 1_039

MOVELOC_UNAVAILABLE_ERROR = 1_040

MOVELOC_EXISTS_BORROW_ERROR = 1_041

BORROWLOC_REFERENCE_ERROR = 1_042

BORROWLOC_UNAVAILABLE_ERROR = 1_043

BORROWLOC_EXISTS_BORROW_ERROR = 1_044

CALL_TYPE_MISMATCH_ERROR = 1_045

CALL_BORROWED_MUTABLE_REFERENCE_ERROR = 1_046

PACK_TYPE_MISMATCH_ERROR = 1_047

UNPACK_TYPE_MISMATCH_ERROR = 1_048

READREF_TYPE_MISMATCH_ERROR = 1_049

READREF_WITHOUT_COPY_ABILITY = 1_050

READREF_EXISTS_MUTABLE_BORROW_ERROR = 1_051

WRITEREF_TYPE_MISMATCH_ERROR = 1_052

WRITEREF_WITHOUT_DROP_ABILITY = 1_053

WRITEREF_EXISTS_BORROW_ERROR = 1_054

WRITEREF_NO_MUTABLE_REFERENCE_ERROR = 1_055

INTEGER_OP_TYPE_MISMATCH_ERROR = 1_056

BOOLEAN_OP_TYPE_MISMATCH_ERROR = 1_057

EQUALITY_OP_TYPE_MISMATCH_ERROR = 1_058

EXISTS_WITHOUT_KEY_ABILITY_OR_BAD_ARGUMENT = 1_059

BORROWGLOBAL_TYPE_MISMATCH_ERROR = 1_060

BORROWGLOBAL_WITHOUT_KEY_ABILITY = 1_061

MOVEFROM_TYPE_MISMATCH_ERROR = 1_062

MOVEFROM_WITHOUT_KEY_ABILITY = 1_063

MOVETO_TYPE_MISMATCH_ERROR = 1_064

MOVETO_WITHOUT_KEY_ABILITY = 1_065

MODULE_ADDRESS_DOES_NOT_MATCH_SENDER = 1_067

NO_MODULE_HANDLES = 1_068

POSITIVE_STACK_SIZE_AT_BLOCK_END = 1_069

MISSING_ACQUIRES_ANNOTATION = 1_070

EXTRANEOUS_ACQUIRES_ANNOTATION = 1_071

DUPLICATE_ACQUIRES_ANNOTATION = 1_072

INVALID_ACQUIRES_ANNOTATION = 1_073

GLOBAL_REFERENCE_ERROR = 1_074

CONSTRAINT_NOT_SATISFIED = 1_075

NUMBER_OF_TYPE_ARGUMENTS_MISMATCH = 1_076

LOOP_IN_INSTANTIATION_GRAPH = 1_077

ZERO_SIZED_STRUCT = 1_080

LINKER_ERROR = 1_081

INVALID_CONSTANT_TYPE = 1_082

MALFORMED_CONSTANT_DATA = 1_083

EMPTY_CODE_UNIT = 1_084

INVALID_LOOP_SPLIT = 1_085

INVALID_LOOP_BREAK = 1_086

INVALID_LOOP_CONTINUE = 1_087

UNSAFE_RET_UNUSED_VALUES_WITHOUT_DROP = 1_088

TOO_MANY_LOCALS = 1_089

GENERIC_MEMBER_OPCODE_MISMATCH = 1_090

FUNCTION_RESOLUTION_FAILURE = 1_091

INVALID_OPERATION_IN_SCRIPT = 1_094

DUPLICATE_MODULE_NAME = 1_095

BACKWARD_INCOMPATIBLE_MODULE_UPDATE = 1_096

CYCLIC_MODULE_DEPENDENCY = 1_097

NUMBER_OF_ARGUMENTS_MISMATCH = 1_098

INVALID_PARAM_TYPE_FOR_DESERIALIZATION = 1_099

FAILED_TO_DESERIALIZE_ARGUMENT = 1_100

NUMBER_OF_SIGNER_ARGUMENTS_MISMATCH = 1_101

CALLED_SCRIPT_VISIBLE_FROM_NON_SCRIPT_VISIBLE = 1_102

EXECUTE_SCRIPT_FUNCTION_CALLED_ON_NON_SCRIPT_VISIBLE = 1_103

INVALID_FRIEND_DECL_WITH_SELF = 1_104

INVALID_FRIEND_DECL_WITH_MODULES_OUTSIDE_ACCOUNT_ADDRESS = 1_105

INVALID_FRIEND_DECL_WITH_MODULES_IN_DEPENDENCIES = 1_106

CYCLIC_MODULE_FRIENDSHIP = 1_107

UNKNOWN_INVARIANT_VIOLATION_ERROR = 2_000

EMPTY_VALUE_STACK = 2_003

PC_OVERFLOW = 2_005

VERIFICATION_ERROR = 2_006

STORAGE_ERROR = 2_008

INTERNAL_TYPE_ERROR = 2_009

EVENT_KEY_MISMATCH = 2_010

UNREACHABLE = 2_011

VM_STARTUP_FAILURE = 2_012

UNEXPECTED_ERROR_FROM_KNOWN_MOVE_FUNCTION = 2_015

VERIFIER_INVARIANT_VIOLATION = 2_016

UNEXPECTED_VERIFIER_ERROR = 2_017

UNEXPECTED_DESERIALIZATION_ERROR = 2_018

FAILED_TO_SERIALIZE_WRITE_SET_CHANGES = 2_019

FAILED_TO_DESERIALIZE_RESOURCE = 2_020

TYPE_RESOLUTION_FAILURE = 2_021

UNKNOWN_BINARY_ERROR = 3_000

MALFORMED = 3_001

BAD_MAGIC = 3_002

UNKNOWN_VERSION = 3_003

UNKNOWN_TABLE_TYPE = 3_004

UNKNOWN_SIGNATURE_TYPE = 3_005

UNKNOWN_SERIALIZED_TYPE = 3_006

UNKNOWN_OPCODE = 3_007

BAD_HEADER_TABLE = 3_008

UNEXPECTED_SIGNATURE_TYPE = 3_009

DUPLICATE_TABLE = 3_010

UNKNOWN_ABILITY = 3_013

UNKNOWN_NATIVE_STRUCT_FLAG = 3_014

BAD_U64 = 3_019

BAD_U128 = 3_020

VALUE_SERIALIZATION_ERROR = 3_022

VALUE_DESERIALIZATION_ERROR = 3_023

CODE_DESERIALIZATION_ERROR = 3_024

INVALID_FLAG_BITS = 3_025

UNKNOWN_RUNTIME_STATUS = 4_000

EXECUTED = 4_001

OUT_OF_GAS = 4_002

RESOURCE_DOES_NOT_EXIST = 4_003

RESOURCE_ALREADY_EXISTS = 4_004

MISSING_DATA = 4_008

DATA_FORMAT_ERROR = 4_009

ABORTED = 4_016

ARITHMETIC_ERROR = 4_017

EXECUTION_STACK_OVERFLOW = 4_020

CALL_STACK_OVERFLOW = 4_021

VM_MAX_TYPE_DEPTH_REACHED = 4_024

VM_MAX_VALUE_DEPTH_REACHED = 4_025

UNKNOWN_STATUS = 18_446_744_073_709_551_615

Implementations

impl StatusCode

pub fn status_type(self) -> StatusType

Return the status type for this status code