Crate diem_types

Modules

access_path

Suppose we have the following data structure in a smart contract:

account_address

account_config

account_state

account_state_blob

block_info

block_metadata

chain_id

committed_block

contract_event

diem_timestamp

epoch_change

epoch_state

event

ledger_info

mempool_status

move_resource

network_address

on_chain_config

proof

reward_distribution_event

serde_helper

term_state

transaction

trusted_state

validator_config

validator_info

validator_signer

validator_verifier

vm_status

waypoint

write_set

For each transaction the VM executes, the VM will output a WriteSet that contains each access path it updates. For each access path, the VM can either give its new value or delete it.

Structs

PeerId

A struct that represents an account address.