Module diem_types::block_metadata

Structs

BlockMetadata

Struct that will be persisted on chain to store the information of the current block.

DiemBlockResource

NewBlockEvent

Statics

NEW_BLOCK_EVENT_PATH

The path to the new block event handle under a DiemBlock::BlockMetadata resource.

Functions

new_block_event_key