Trait diem_types::move_resource::MoveStorage

pub trait MoveStorage {
    fn batch_fetch_resources(
        &self, 
        access_paths: Vec<AccessPath>
    ) -> Result<Vec<Vec<u8>>>;

    fn batch_fetch_resources_by_version(
        &self, 
        access_paths: Vec<AccessPath>, 
        version: Version
    ) -> Result<Vec<Vec<u8>>>;

    fn fetch_synced_version(&self) -> Result<Version>;
}

Required methods

fn batch_fetch_resources(
&self,
access_paths: Vec<AccessPath>
) -> Result<Vec<Vec<u8>>>

Returns a vector of Move resources as serialized byte array Order of resources returned matches the order of access_path

fn batch_fetch_resources_by_version(
&self,
access_paths: Vec<AccessPath>,
version: Version
) -> Result<Vec<Vec<u8>>>

Returns a vector of Move resources as serialized byte array from a specified version of the database Order of resources returned matches the order of access_path

fn fetch_synced_version(&self) -> Result<Version>

Get the version on the latest transaction info.

Loading content...