Trait diem_types::on_chain_config::OnChainConfig

pub trait OnChainConfig: Send + Sync + DeserializeOwned {
    const IDENTIFIER: &'static str;
    const ADDRESS: &'static str = CONFIG_ADDRESS_STR;
    const CONFIG_ID: ConfigID = _;

    fn deserialize_default_impl(bytes: &[u8]) -> Result<Self> { ... }
    fn deserialize_into_config(bytes: &[u8]) -> Result<Self> { ... }
    fn fetch_config<T>(storage: &T) -> Option<Self>
    where
        T: ConfigStorage,
    { ... }
}

Trait to be implemented by a Rust struct representation of an on-chain config that is stored in storage as a serialized byte array

Required Associated Constants

const IDENTIFIER: &'static str

Provided Associated Constants

const ADDRESS: &'static str = CONFIG_ADDRESS_STR

const CONFIG_ID: ConfigID = _

Provided Methods

fn deserialize_default_impl(bytes: &[u8]) -> Result<Self>

fn deserialize_into_config(bytes: &[u8]) -> Result<Self>

fn fetch_config<T>(storage: &T) -> Option<Self>where
T: ConfigStorage,