diem_types::proof

Module accumulator

This module implements an in-memory Merkle Accumulator that is similar to what we use in storage. This accumulator will only store a small portion of the tree – for any subtree that is full, we store only the root. Also we only store the frozen nodes, therefore this structure will always store up to Log(n) number of nodes, where n is the total number of leaves in the tree.

This accumulator is immutable once constructed. If we append new leaves to the tree we will obtain a new accumulator instance and the old one remains unchanged.

Structs

InMemoryAccumulator

The Accumulator implementation.